🚚 Relocate App.tsx file
